Refactor issue service to use the new issue mock and add lombok.
