+++ issue/src/main/java/de/upb/crypto/incentive/services/issue/IssueController.java
@@ -1,8 +1,8 @@
 package de.upb.crypto.incentive.services.issue;
 
-import de.upb.crypto.incentive.procotols.issue.IssueHelper;
-import de.upb.crypto.incentive.procotols.issue.IssueRequest;
-import de.upb.crypto.incentive.procotols.issue.IssueResponse;
+import de.upb.crypto.incentive.protocols.issue.IssueRequest;
+import de.upb.crypto.incentive.protocols.issue.IssueResponse;
+import de.upb.crypto.incentive.protocols.model.Token;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
@@ -11,12 +11,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Locale;
-
 @RestController
 public class IssueController {
-
-    IssueHelper issueHelper = new IssueHelper();
 
     @GetMapping("/issue")
     @ApiOperation(value = "Issuing protocol", notes = "Issue a new incentive token.", response = IssueResponse.class)
@@ -24,11 +20,8 @@ public class IssueController {
             @ApiResponse(code = 200, message = "Success", response = IssueResponse.class),
             @ApiResponse(code = 403, message = "Invalid Issuing Request", response = String.class)
     })
-    public ResponseEntity<IssueResponse> greeting(@Validated @ModelAttribute IssueRequest request) throws IncentiveException {
-        if (request.equals(issueHelper.validRequest)) {
-            return new ResponseEntity<>(issueHelper.validResponse, HttpStatus.OK);
-        }
-        throw new IncentiveException();
+    public ResponseEntity<IssueResponse> greeting(@Validated IssueRequest request) {
+        return new ResponseEntity<>(new IssueResponse(request.getId(), new Token(0)), HttpStatus.OK);
     }
 
     @ResponseStatus(HttpStatus.FORBIDDEN)
