+++ issue/src/main/java/de/upb/crypto/incentive/services/issue/SwaggerConfig.java
@@ -1,0 +1,21 @@
+package de.upb.crypto.incentive.services.issue;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig {
+
+    @Bean
+    public Docket apiDocket() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .groupName("issue-api")
+                .select()
+                .build();
+    }
+}
+
